% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue_dependencies.R
\name{catalogue_dependencies}
\alias{catalogue_dependencies}
\title{Look for library calls in a directory of Rmds, install
any packages you don't have, then add packages to a
DESCRIPTION}
\usage{
catalogue_dependencies(path = "./presentations", install = TRUE)
}
\arguments{
\item{path}{Directory of files to scan}

\item{install}{Should installs happen}
}
\value{
Currently a lot of gunk!
}
\description{
Look for library calls in a directory of Rmds, install
any packages you don't have, then add packages to a
DESCRIPTION
}
\examples{
catalogue_dependencies(install=FALSE)
}
