% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_htmlfiles.R
\name{generate_htmlfiles}
\alias{generate_htmlfiles}
\title{Take a directory of markdown files and render to
HTML any files not previously rendered, or with
last modifed dates greater than the HTML output}
\usage{
generate_htmlfiles(path = "./content/", dest = "./static/",
  template = "default.revealjs")
}
\arguments{
\item{path}{Directory of mds}

\item{dest}{Directory of html}
}
\value{
Render outputs
}
\description{
Take a directory of markdown files and render to
HTML any files not previously rendered, or with
last modifed dates greater than the HTML output
}
\examples{
\dontrun{
generate_htmlfiles()
}
}
