% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mdfiles.R
\name{generate_mdfiles}
\alias{generate_mdfiles}
\title{Take a directory of rmarkdown files and render to
markdown any files not previously rendered, or with
last modifed dates greater than the markdown output}
\usage{
generate_mdfiles(path = "./presentations/", dest = "./content/")
}
\arguments{
\item{path}{Directory of Rmds}

\item{dest}{Directory of mds}
}
\value{
Render outputs
}
\description{
Take a directory of rmarkdown files and render to
markdown any files not previously rendered, or with
last modifed dates greater than the markdown output
}
\examples{
\dontrun{
generate_mdfiles()
}
}
