% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_presentationjs.R
\name{install_presentationjs}
\alias{install_presentationjs}
\title{Grab files needed to use a presentation library}
\usage{
install_presentationjs(tmp_dir = tempdir(),
  pres_lib_dest = "./static/revealjs",
  pres_lib_loc = "https://github.com/hakimel/reveal.js/archive/3.6.0.zip",
  pres_lib_template = "https://raw.githubusercontent.com/jgm/pandoc-templates/master/default.revealjs")
}
\arguments{
\item{tmp_dir}{Specific directory to download stuff to}

\item{pres_lib_dest}{The directory name to deploy js library to}

\item{pres_lib_loc}{The source of the js library}

\item{pres_lib_template}{The relevant template for converting md to html}
}
\value{
Lots of gunk right now
}
\description{
Grab files needed to use a presentation library
}
\examples{
\dontrun{
install_presentationjs()
}
}
