% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presdown-package.r
\docType{package}
\name{presdown}
\alias{presdown}
\alias{presdown-package}
\title{Build an online directory of slides in a way that's
friendly to you and continuous integration!
You use presdown to generate slides that haven't
been generated before or have out of date outputs,
incorporating dependencies listed in the DESCRIPTION
file and you can even grab the list based on a
directory of presentations.}
\description{
Build an online directory of slides in a way that's
friendly to you and continuous integration!
You use presdown to generate slides that haven't
been generated before or have out of date outputs,
incorporating dependencies listed in the DESCRIPTION
file and you can even grab the list based on a
directory of presentations.
}
