% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\name{setup_dir}
\alias{setup_dir}
\title{Generate improtant directories according to a
stagtic site generators preferences.}
\usage{
setup_dir(path = ".", convention = c("hugo", "jekyll"))
}
\arguments{
\item{path}{Directory in which directories should be created}

\item{convention}{Static site generator conventions to use}
}
\value{
Boolean indicating success
}
\description{
Generate improtant directories according to a
stagtic site generators preferences.
}
\examples{
setup_dir(convention="hugo")
}
